Server.Dockerfile: moved the location of running deploy base
Server.Dockerfile
- running deploy base
- copying some files to /noobaa_init_files/

noobaa.spec
- removed the run of deploy base

install_noobaa.sh
- no longer yum installing the rpm

supervisord.orig
- In docker running the noobaa_init.sh from a diffrent path

fix_server_plat.sh
- Add get_env_in_docker function that will return the .env that we
copied out in the Server.Dockerfile

setup_mongo_ssl.sh:
- Divided into functions

fix_mongo_ssl.sh:
- minor fix for debug

noobaa_core.yaml:
- Add the ability to set the container in dbg

+++ src/deploy/NVA_build/noobaa_init.sh
@@ -86,7 +86,8 @@ fix_non_root_user() {
 extract_noobaa_in_docker() {
   local file
   local files=(deploy_base.sh noobaa-NVA.tar.gz noobaa.rpm)
-  if [ "${container}" == "docker" ] ; then
+  local noobaa_core_path="/root/node_modules/noobaa-core/"
+  if [ "${container}" == "docker" ] && [ ! -d ${noobaa_core_path} ] ; then
     cd /tmp/
     rpm2cpio noobaa.rpm | cpio -idmv
     rm -rf /tmp/deploy_base.sh
@@ -101,7 +102,8 @@ extract_noobaa_in_docker() {
 }
 
 run_kube_pv_chown() {
-  # change ownership and permissions of /data and /log. assuming that uid is not changed between reboots
+  # change ownership and permissions of /data and /log. 
+  # assuming that uid is not changed between reboots.
   local path="/root/node_modules/noobaa-core/build/Release/"
   if [ "${container}" == "docker" ] ; then
       path="/noobaa_init_files/"
